@@ -1,58 +1,58 @@
 {
-    "name": "weather-ui",
-    "version": "1.0.0",
-    "engines": {
-        "npm": ">=10.1.0 <11.0.0",
-        "node": ">=20.9.0 <21.0.0"
-    },
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test",
-        "lint": "ng lint",
-        "test-headless": "ng test --watch=false --browsers=ChromeHeadless"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^17.3.12",
-        "@angular/common": "^17.3.12",
-        "@angular/compiler": "^17.3.12",
-        "@angular/core": "^17.3.12",
-        "@angular/forms": "^17.3.12",
-        "@angular/platform-browser": "^17.3.12",
-        "@angular/platform-browser-dynamic": "^17.3.12",
-        "@angular/router": "^17.3.12",
-        "chart.js": "^4.4.4",
-        "rxjs": "~7.8.1",
-        "tslib": "^2.7.0",
-        "zone.js": "^0.14.10"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^17.3.10",
-        "@angular-eslint/builder": "17.5.3",
-        "@angular-eslint/eslint-plugin": "17.5.3",
-        "@angular-eslint/eslint-plugin-template": "17.5.3",
-        "@angular-eslint/schematics": "17.5.3",
-        "@angular-eslint/template-parser": "17.5.3",
-        "@angular/cli": "^17.3.10",
-        "@angular/compiler-cli": "^17.3.12",
-        "@types/jasmine": "~5.1.4",
-        "@typescript-eslint/eslint-plugin": "7.18.0",
-        "@typescript-eslint/parser": "7.18.0",
-        "autoprefixer": "^10.4.20",
-        "daisyui": "^4.12.10",
-        "eslint": "^8.57.1",
-        "jasmine-core": "~5.3.0",
-        "karma": "~6.4.4",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.1",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.1.0",
-        "postcss": "^8.4.47",
-        "prettier": "^3.3.3",
-        "tailwindcss": "^3.4.13",
-        "typescript": "~5.4.5"
-    }
+  "name": "weather-ui",
+  "version": "1.0.0",
+  "engines": {
+    "npm": ">=10.1.0 <11.0.0",
+    "node": ">=20.9.0 <21.0.0"
+  },
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "lint": "ng lint",
+    "test-headless": "ng test --watch=false --browsers=ChromeHeadless"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^18.2.6",
+    "@angular/common": "^18.2.6",
+    "@angular/compiler": "^18.2.6",
+    "@angular/core": "^18.2.6",
+    "@angular/forms": "^18.2.6",
+    "@angular/platform-browser": "^18.2.6",
+    "@angular/platform-browser-dynamic": "^18.2.6",
+    "@angular/router": "^18.2.6",
+    "chart.js": "^4.4.4",
+    "rxjs": "~7.8.1",
+    "tslib": "^2.7.0",
+    "zone.js": "^0.14.10"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^18.2.6",
+    "@angular-eslint/builder": "18.3.1",
+    "@angular-eslint/eslint-plugin": "18.3.1",
+    "@angular-eslint/eslint-plugin-template": "18.3.1",
+    "@angular-eslint/schematics": "18.3.1",
+    "@angular-eslint/template-parser": "18.3.1",
+    "@angular/cli": "^18.2.6",
+    "@angular/compiler-cli": "^18.2.6",
+    "@types/jasmine": "~5.1.4",
+    "@typescript-eslint/eslint-plugin": "7.18.0",
+    "@typescript-eslint/parser": "7.18.0",
+    "autoprefixer": "^10.4.20",
+    "daisyui": "^4.12.10",
+    "eslint": "^8.57.1",
+    "jasmine-core": "~5.3.0",
+    "karma": "~6.4.4",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.1",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "postcss": "^8.4.47",
+    "prettier": "^3.3.3",
+    "tailwindcss": "^3.4.13",
+    "typescript": "~5.4.5"
+  }
 }
